#59 refactor: req, res package 생성
