+++ src/main/kotlin/com/psr/psr/user/dto/request/UserEidReq.kt
@@ -1,4 +1,4 @@
-package com.psr.psr.user.dto
+package com.psr.psr.user.dto.request
 
 import com.psr.psr.user.dto.eidReq.BusinessListReq
 import com.psr.psr.user.dto.eidReq.Business
