+++ src/main/kotlin/com/psr/psr/user/dto/request/UserInterestListReq.kt
@@ -1,4 +1,4 @@
-package com.psr.psr.user.dto
+package com.psr.psr.user.dto.request
 
 data class UserInterestListReq (
     val interestList: List<UserInterestReq>
